Fix simi )"

+++ plugins/ttp.js
@@ -7,7 +7,7 @@ const { MessageType } = require('@adiwajshing/baileys')
 
 let handler  = async (m, { conn, text }) => {
   if (text) {
-    let res = await fetch('https://api.xteam.xyz/ttp?file&text=' + encodeURIComponent(text))
+    let res = await fetch(global.API('xteam', '/ttp', { file: '', text }))
     let img = await res.buffer()
     if (!img) throw img
     let stiker = await sticker(img)
@@ -56,9 +56,11 @@ function sticker(img, url) {
         spawn('convert', [png, out])
         .on('error', reject)
         .on('close', () => {
-          fs.unlinkSync(png)
-          resolve(fs.readFileSync(out))
-          fs.unlinkSync(out)
+          try {
+            fs.unlinkSync(png)
+            resolve(fs.readFileSync(out))
+            if (fs.existSync(out)) fs.unlinkSync(out)
+          } catch (e) { reject(e) }
         })
       })
     } catch (e) {
